* StdLib.dsc was changed to always build the sockets code. * Fix errors detected by the GCC compiler (pointer conversions and format mismatches). * Worked around a GCC flow analysis issue in base64.c where the compiler is convinced that there is a path through the code where "input" is used but not initialized. * Added EFIAPI to the file system operation routines for sockets to match the API declaration.
Note for GCC, the meaning of "l" and "L" are different in printf and sscanf format strings.  The lower case "l" indicates a 32-bit value where the capital "L" indicates a 64-bit value.  The native (default) integer size may be represented by a size letter of "n".



git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@13023 6f19259b-4bc3-4df7-8a09-765794883524

+++ StdLib/BsdSocketLib/inet_net_ntop.c
@@ -126,7 +126,7 @@ inet_net_ntop_ipv4(
 		if (dst != odst)
 			*dst++ = '.';
 		m = ((1 << b) - 1) << (8 - b);
-		dst += SPRINTF((dst, "%u", *src & m));
+		dst += SPRINTF((dst, "%u", ((unsigned int)(*src & m))));
 		size -= (size_t)(dst - t);
 	}
